+++ oauth2_provider/views/application.py
@@ -3,7 +3,22 @@ from django.forms.models import modelform_factory
 from django.urls import reverse_lazy
 from django.views.generic import CreateView, DeleteView, DetailView, ListView, UpdateView
 
+from ..forms import ApplicationForm
 from ..models import get_application_model
+
+
+APPLICATION_FIELDS = (
+    "name",
+    "client_id",
+    "client_secret",
+    "hash_client_secret",
+    "client_type",
+    "authorization_grant_type",
+    "redirect_uris",
+    "post_logout_redirect_uris",
+    "allowed_origins",
+    "algorithm",
+)
 
 
 class ApplicationOwnerIsUserMixin(LoginRequiredMixin):
@@ -25,24 +40,7 @@ class ApplicationRegistration(LoginRequiredMixin, CreateView):
     template_name = "oauth2_provider/application_registration_form.html"
 
     def get_form_class(self):
-        """
-        Returns the form class for the application model
-        """
-        return modelform_factory(
-            get_application_model(),
-            fields=(
-                "name",
-                "client_id",
-                "client_secret",
-                "hash_client_secret",
-                "client_type",
-                "authorization_grant_type",
-                "redirect_uris",
-                "post_logout_redirect_uris",
-                "allowed_origins",
-                "algorithm",
-            ),
-        )
+        return modelform_factory(get_application_model(), form=ApplicationForm, fields=APPLICATION_FIELDS)
 
     def form_valid(self, form):
         form.instance.user = self.request.user
@@ -86,21 +84,4 @@ class ApplicationUpdate(ApplicationOwnerIsUserMixin, UpdateView):
     template_name = "oauth2_provider/application_form.html"
 
     def get_form_class(self):
-        """
-        Returns the form class for the application model
-        """
-        return modelform_factory(
-            get_application_model(),
-            fields=(
-                "name",
-                "client_id",
-                "client_secret",
-                "hash_client_secret",
-                "client_type",
-                "authorization_grant_type",
-                "redirect_uris",
-                "post_logout_redirect_uris",
-                "allowed_origins",
-                "algorithm",
-            ),
-        )
+        return modelform_factory(get_application_model(), form=ApplicationForm, fields=APPLICATION_FIELDS)